MWPW-167326: CaaS BulkPublisher v2 release  (#3743)
* Revert "MWPW-140452 - Icon authoring in milo using the federal repo a… (#3357)

Revert "MWPW-140452 - Icon authoring in milo using the federal repo and individual SVG assets (#3259)"

This reverts commit 81a5770cecc2a17f07efe5229c65b71d2dbb9bb4.

* [Release] Stage to Main (#3497)

MWPW-165774 [Mobile-GNAV] page is not scrollable in live page (#3495)

* check for new nav when disabling ios scroll

* shortened the check from the previous commit


* Promotes beta version to default URL

* MWPW-167326: adds download CSV option and code cleanup

* MWPW-167326: fixes eslint errors

* MWPW-167326: fixes eslint errors II

* MWPW-167326: fixes eslint errors III

* MWPW-167326: fixes eslint errors IV

* MWPW-167326: fixes eslint errors V

---------
